week-1/requirements.txt
vosk
openai-whisper
jiwer
sounddevice
torch
